Moved parsing string literal
