+++ src/YINIVisitor.ts
@@ -21,6 +21,7 @@ import {
     YiniContext,
 } from './grammar/YiniParser.js'
 import YiniParserVisitor from './grammar/YiniParserVisitor'
+import parseStringLiteral from './literal-parsers/parseStringLiteral'
 import { debugPrint, isDebug } from './utils/general'
 
 // const isDebug = !!process.env.IS_DEBUG
@@ -301,47 +302,47 @@ export default class YINIVisitor<IResult> extends YiniParserVisitor<IResult> {
         debugPrint('-> Entered visitString_literal(..)')
 
         // @todo TODO: Move the parsing of raw into a file into literal-parsers/
-        let raw = ctx.getText()
+        const raw = ctx.getText()
         debugPrint('raw = >>>' + raw + '<<<')
 
-        /*
-            Extracts an optional prefix (C, c, H, or h) and identifies whether
-            the string is triple-quoted, double-quoted, or single-quoted.
-        */
-        const prefixMatch = raw.match(/^(C|c|H|h|R|r)?("""|"|')/)
-        debugPrint('prefixMatch:')
-        if (isDebug()) {
-            console.debug(prefixMatch)
-        }
+        // /*
+        //     Extracts an optional prefix (C, c, H, or h) and identifies whether
+        //     the string is triple-quoted, double-quoted, or single-quoted.
+        // */
+        // const prefixMatch = raw.match(/^(C|c|H|h|R|r)?("""|"|')/)
+        // debugPrint('prefixMatch:')
+        // if (isDebug()) {
+        //     console.debug(prefixMatch)
+        // }
 
-        let prefix = prefixMatch ? prefixMatch[1]?.toUpperCase() : ''
-        debugPrint('          prefix = ' + prefix)
+        // let prefix = prefixMatch ? prefixMatch[1]?.toUpperCase() : ''
+        // debugPrint('          prefix = ' + prefix)
 
-        let quoteType = prefixMatch ? prefixMatch[2] : ''
-        debugPrint('       quoteType = ' + quoteType)
-        debugPrint('quoteType.length = ' + quoteType.length)
+        // let quoteType = prefixMatch ? prefixMatch[2] : ''
+        // debugPrint('       quoteType = ' + quoteType)
+        // debugPrint('quoteType.length = ' + quoteType.length)
 
-        // Extracts the substring after removing the initial prefix (if any)
-        // and quotes at the start (prefix.length + quoteType.length) and the
-        // quotes at the end (-quoteType.length).
-        let inner = raw.slice(
-            (prefix?.length || 0) + quoteType.length,
-            -quoteType.length,
-        )
-        debugPrint('inner (raw) = ' + inner)
+        // // Extracts the substring after removing the initial prefix (if any)
+        // // and quotes at the start (prefix.length + quoteType.length) and the
+        // // quotes at the end (-quoteType.length).
+        // let inner = raw.slice(
+        //     (prefix?.length || 0) + quoteType.length,
+        //     -quoteType.length,
+        // )
+        // debugPrint('inner (raw) = ' + inner)
 
-        if (prefix === 'C') {
-            inner = inner
-                .replace(/\\n/g, '\n')
-                .replace(/\\t/g, '\t')
-                .replace(/\\r/g, '\r')
-        } else if (prefix === 'H') {
-            inner = inner.replace(/[\s\n\r]+/g, ' ').trim()
-        }
+        // if (prefix === 'C') {
+        //     inner = inner
+        //         .replace(/\\n/g, '\n')
+        //         .replace(/\\t/g, '\t')
+        //         .replace(/\\r/g, '\r')
+        // } else if (prefix === 'H') {
+        //     inner = inner.replace(/[\s\n\r]+/g, ' ').trim()
+        // }
 
-        debugPrint('inner (reformat) = ' + inner)
+        // debugPrint('inner (reformat) = ' + inner)
 
-        return { type: 'String', value: inner } as IResult
+        return { type: 'String', value: parseStringLiteral(raw) } as IResult
     }
 
     /**
